+++ bfabric_app_runner/src/bfabric_app_runner/dispatch/dispatch_individual_resources.py
@@ -96,13 +96,14 @@ class DispatchIndividualResources:
             msg = f"Dataset with id {definition.execution.dataset} not found"
             raise ValueError(msg)
         dataset_df = dataset.to_polars()
-        resources = self._client.reader.read_ids(Resource, dataset_df[config.resource_column].unique().to_list())
-        resources_by_id = {uri.components.entity_id: resource for uri, resource in resources.items()}
+        resources_by_id = self._client.reader.read_ids(
+            Resource, dataset_df[config.resource_column].unique().to_list()
+        ).by_id
         paths = []
         for row in dataset_df.iter_rows(named=True):
             resource_id = int(row[config.resource_column])
             row_params = {name: row[dataset_name] for dataset_name, name in config.param_columns}
-            resource = resources_by_id[resource_id]
+            resource = resources_by_id.get(resource_id)
             if resource is None:
                 msg = f"Resource with id {resource_id} not found"
                 raise ValueError(msg)