+++ bfabric_app_runner/src/bfabric_app_runner/inputs/resolve/_resolve_bfabric_resource_dataset_specs.py
@@ -7,7 +7,6 @@ from typing import TYPE_CHECKING
 import polars as pl
 
 from bfabric.entities import Dataset, Resource
-from bfabric.entities.core.reader_utils import present_entities
 from bfabric_app_runner.inputs._filter_files import filter_dataframe
 from bfabric_app_runner.inputs.resolve._common import get_ssh_file_source
 from bfabric_app_runner.inputs.resolve.resolved_inputs import ResolvedFile, ResolvedStaticFile
@@ -98,7 +97,7 @@ class ResolveBfabricResourceDatasetSpecs:
                 "tmp_resource_relative_path": r.storage_relative_path,
                 "tmp_resource_source": get_ssh_file_source(r),
             }
-            for r in present_entities(self._reader.read_ids(Resource, resource_ids))
+            for r in self._reader.read_ids(Resource, resource_ids).present
         ]
 
         # Merge